feat(sdk): move things around
