+++ sdk/src/utils.ts
@@ -2,16 +2,6 @@
 // Modifications Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import {
-    hasAvailableClaims,
-    isCouponExpired,
-    isCouponValidForAddress,
-    isCouponValidForNameLength,
-    isCouponValidForNameYears,
-    isValidCouponPercentage,
-} from './coupons';
-import { Coupon } from './types';
-
 const LABEL_REGEX = /(?!-)[a-z0-9-]{0,62}[a-z0-9]/;
 const SUBNAME_REGEX = /^(?!-)[a-z0-9-]{1,62}[a-z0-9]$/;
 const PATH_REGEX = new RegExp(`(?:${LABEL_REGEX.source}(?:\\.${LABEL_REGEX.source})*)`);
@@ -124,69 +114,4 @@ export function validateIotaName(
         }
     }
     return null;
-}
-
-export function validateCoupon(coupon: Coupon, years: number, length: number, address?: string) {
-    try {
-        hasAvailableClaims(coupon.rules);
-        isCouponValidForNameYears(coupon.rules, years);
-        isValidCouponPercentage(coupon.amount);
-        isCouponValidForNameLength(coupon.rules, length);
-        isCouponValidForAddress(coupon.rules, address);
-        isCouponExpired(coupon.rules);
-    } catch (error: unknown) {
-        throw new Error(
-            `Coupon '${coupon.couponCode}' validation failed: ${(error as Error)?.message}`,
-        );
-    }
-}
-
-export function validateCoupons(
-    coupons: Coupon[],
-    years: number,
-    length: number,
-    address?: string,
-) {
-    for (const coupon of coupons) {
-        validateCoupon(coupon, years, length, address);
-    }
-}
-
-export function applyCouponsToPrice(coupons: Coupon[], initialPrice: number): number {
-    if (!coupons || coupons.length === 0) {
-        return initialPrice;
-    }
-
-    let price = initialPrice;
-
-    for (const coupon of coupons) {
-        if (!coupon.rules.can_stack && coupons.length > 1) {
-            throw new Error('Coupons provided cannot be stacked');
-        }
-
-        price = applyCouponToPrice(price, coupon);
-    }
-
-    return price;
-}
-
-export function applyCouponToPrice(price: number, coupon?: Coupon): number {
-    if (!coupon) {
-        return price;
-    }
-
-    const couponAmount = Number(coupon.amount);
-
-    // 0 => percentage off
-    // 1 => fixed amount off,
-    if (coupon.kind === 0) {
-        let discountAmount = (price * couponAmount) / 100;
-        return price - discountAmount;
-    } else if (coupon.kind === 1) {
-        const discountedAmount = price - couponAmount;
-
-        return discountedAmount < 0 ? 0 : discountedAmount;
-    } else {
-        throw new Error(`Unknown coupon kind: ${coupon.kind}`);
-    }
 }
